main: create KeyStates from ScrollEvents
To be able to bind things to scroll events, they have to be converted
to KeyEvents with a pressed/released state.

Scroll events only send a scroll increment, repeatedly.  This has to
be converted into (synthetic) press and release events.  To this end
some additional book-keeping on the key states, event times, ... is
added.